6755277: All initial changes to jdk7 to create openjdk 6 build 0 Summary: Final b00 state (as defined by the source bundle) Reviewed-by: darcy

+++ src/share/classes/sun/net/ProgressEvent.java
@@ -30,83 +30,86 @@ import java.net.URL;
 /**
  * ProgressEvent represents an progress event in monitering network input stream.
  *
- * @author Stanley Man-Kit Ho
+ * @author Stanley Man-Kit Ho 
  */
-public class ProgressEvent extends EventObject  {
+public class ProgressEvent extends EventObject	{
     // URL of the stream
     private URL url;
     // content type of the stream
     private String contentType;
     // method associated with URL
     private String method;
-    // bytes read
-    private long progress;
-    // bytes expected
-    private long expected;
+    // bytes read 
+    private int progress;
+    // bytes expected 
+    private int expected;
     // the last thing to happen
     private ProgressSource.State state;
-
+ 
     /**
      * Construct a ProgressEvent object.
      */
-    public ProgressEvent(ProgressSource source, URL url, String method, String contentType, ProgressSource.State state, long progress, long expected) {
-        super(source);
-        this.url = url;
-        this.method = method;
-        this.contentType = contentType;
-        this.progress = progress;
-        this.expected = expected;
-        this.state = state;
+    public ProgressEvent(ProgressSource source, URL url, String method, String contentType, ProgressSource.State state, int progress, int expected) {
+	super(source);
+	this.url = url;
+	this.method = method;
+	this.contentType = contentType;
+	this.progress = progress;
+	this.expected = expected;
+	this.state = state;
     }
 
     /**
      * Return URL related to the progress.
      */
-    public URL getURL()
+    public URL getURL()	
     {
-        return url;
+	return url;
     }
 
     /**
      * Return method associated with URL.
      */
-    public String getMethod()
+    public String getMethod()	
     {
-        return method;
+	return method;
     }
 
-    /**
+    /** 
      * Return content type of the URL.
      */
-    public String getContentType()
+    public String getContentType()  
     {
-        return contentType;
-    }
+	return contentType;
+    }    
 
     /**
      * Return current progress value.
      */
-    public long getProgress()
+    public int getProgress()	
     {
-        return progress;
+	return progress;
     }
-
-    /**
+    
+    /** 
      * Return expected maximum progress value; -1 if expected is unknown.
      */
-    public long getExpected() {
-        return expected;
+    public int getExpected() {
+	return expected;
     }
-
+    
     /**
      * Return state.
      */
     public ProgressSource.State getState() {
-        return state;
+	return state;
     }
-
-    public String toString()    {
-        return getClass().getName() + "[url=" + url + ", method=" + method + ", state=" + state
-             + ", content-type=" + contentType + ", progress=" + progress + ", expected=" + expected + "]";
+    
+    public String toString()	{
+	return getClass().getName() + "[url=" + url + ", method=" + method + ", state=" + state 
+	     + ", content-type=" + contentType + ", progress=" + progress + ", expected=" + expected + "]";
     }
 }
+
+
+
